extract classes to own file
