+++ tests/openscadtestframework/__init__.py
@@ -1,5 +1,5 @@
 # __init__.py
 
 __all__ = ["IntegrationTest", "Module", "ModuleTest", "Cube", "Square"]
-from .openscadtestframework import IntegrationTest, Module, ModuleTest
-from .basic_modules import Cube, Square
+from .tests import IntegrationTest, ModuleTest
+from .modules import Module, Cube, Square
